# docs/openapi_spec.yaml
openapi: 3.0.0
info:
  description: APIs for Vending Machine Service
  version: "1.0.0"
  title: Vending Machine Service
  contact:
    email: vengateshs@gmail.com
servers:
  - url: 'http://127.0.0.1'
...
paths:
...
  /products:
    post:
      tags:
        - Products
      summary: Create a new product
      description: POST can be called only by the seller user who created the product
      operationId: createProduct
      security:
        - basicAuth: []      
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
        required: true      
      responses:
        "200":
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "503":
          description: Service Unavilable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
    get:
      tags:
        - Products
      summary: List all the registered products
      description: GET can be called by anyone
      operationId: getAllProducts
      security:
        - basicAuth: []      
      parameters:
      - description: The number of users to skip before starting to collect the result set
        in: query
        name: offset
        schema:
          type: integer
      - description: The numbers of items to return
        in: query
        name: limit
        schema:
          type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "503":
          description: Service Unavilable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get detailed information of a registered product
      description: Get detailed information of a registered product by productId.
      operationId: getProduct
      security:
        - basicAuth: []
      parameters:
      - in: path
        name: productId
        required: true
        schema:
          $ref: '#/components/schemas/Product/properties/id'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "503":
          description: Service Unavilable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Products
      summary: Delete a registered product
      description: Delete a registered product by productId. DELETE can be called only by the seller user who created the product
      operationId: deleteProduct
      security:
        - basicAuth: []
      parameters:
      - in: path
        name: productId
        required: true
        schema:
          $ref: '#/components/schemas/Product/properties/id'
      responses:
        "200":
          description: Product deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "503":
          description: Service Unavilable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Products
      summary: Update a registered product
      description: Update a registered product by productId. PATCH can be called only by the seller user who created the product
      operationId: updateProduct
      security:
        - basicAuth: []
      parameters:
      - in: path
        name: productId
        required: true
        schema:
          $ref: '#/components/schemas/Product/properties/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
        required: true
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Entity Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "406":
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
        "503":
          description: Service Unavilable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ProductCreate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: double
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        createdBy:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: double
    ProductList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Product'
          type: array
        limit:
          type: integer
        offset:
          type: integer
        total:
          description: Total number of products
          type: integer
      type: object
    ErrorResponse:
      properties:
        error:
          description: A user friendly error message
          example: An example error message.
          type: string
        errorCode:
          description: A machine-friendly identifier for the error response
          type: string
        traceId:
          description: A unique identifier for the request
          type: string
      required:
      - error
      - errorCode
      - traceId
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic